lostoderos-back/src/test/java/co/edu/uniandes/csw/lostoderos/test/persistence/PersonaNaturalPersistenceTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package co.edu.uniandes.csw.lostoderos.test.persistence;

import co.edu.uniandes.csw.lostoderos.entities.PersonaNaturalEntity;
import co.edu.uniandes.csw.lostoderos.persistence.PersonaNaturalPersistence;
import javax.inject.Inject;
import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import javax.transaction.UserTransaction;
import org.jboss.arquillian.container.test.api.Deployment;
import org.jboss.arquillian.junit.Arquillian;
import org.jboss.shrinkwrap.api.ShrinkWrap;
import org.jboss.shrinkwrap.api.spec.JavaArchive;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import uk.co.jemos.podam.api.PodamFactory;
import uk.co.jemos.podam.api.PodamFactoryImpl;

/**
 *
 * @author na.morenoe
 */
@RunWith(Arquillian.class)
public class PersonaNaturalPersistenceTest {
    
        /**
     *
     * @return Devuelve el jar que Arquillian va a desplegar en el Glassfish
     * embebido. El jar contiene las clases de Editorial, el descriptor de la
     * base de datos y el archivo beans.xml para resolver la inyección de
     * dependencias.
     */
    @Deployment
    public static JavaArchive createDeployment() {
        return ShrinkWrap.create(JavaArchive.class)
                .addPackage(PersonaNaturalEntity.class.getPackage())
                .addPackage(PersonaNaturalPersistence.class.getPackage())
                .addAsManifestResource("META-INF/persistence.xml", "persistence.xml")
                .addAsManifestResource("META-INF/beans.xml", "beans.xml");
    }
    
        /**
     * Inyección de la dependencia a la clase EditorialPersistence cuyos métodos
     * se van a probar.
     */
    @Inject
    private PersonaNaturalPersistence personanaturalPersistence;

    /**
     * Contexto de Persistencia que se va a utilizar para acceder a la Base de
     * datos por fuera de los métodos que se están probando.
     */
    @PersistenceContext
    private EntityManager em;

    /**
     * Variable para martcar las transacciones del em anterior cuando se
     * crean/borran datos para las pruebas.
     */
    @Inject
    UserTransaction utx;
    
    @Test
    public void createEditorialTest() {
        PodamFactory factory = new PodamFactoryImpl();
        PersonaNaturalEntity newEntity = factory.manufacturePojo(PersonaNaturalEntity.class);
        PersonaNaturalEntity result = personanaturalPersistence.create(newEntity);

        Assert.assertNotNull(result);

        PersonaNaturalEntity entity = em.find(PersonaNaturalEntity.class, result.getId());

        Assert.assertEquals(newEntity.getName(), entity.getName());
    }
    
}
